norm ok
