+++ src/philosophers_behaiver.c
@@ -6,49 +6,11 @@
 /*   By: kotkobay <kotkobay@student.42tokyo.jp>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/10/24 12:52:55 by kotkobay          #+#    #+#             */
-/*   Updated: 2024/11/27 11:18:37 by kotkobay         ###   ########.fr       */
+/*   Updated: 2024/11/27 11:30:36 by kotkobay         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../inc/philo.h"
-
-void	check_elapsed_time(t_philosophers *philo)
-{
-	long	elapsed_ms;
-
-	if (gettimeofday(&philo->now, NULL) != 0)
-		exit_with_message("Error: gettimeofday failed");
-	elapsed_ms = (philo->now.tv_sec - philo->start.tv_sec) * 1000;
-	elapsed_ms += (philo->now.tv_usec - philo->start.tv_usec) / 1000;
-	if (elapsed_ms >= philo->argument->time_to_die)
-		handle_death(philo);
-}
-
-void	handle_death(t_philosophers *philo)
-{
-	pthread_mutex_lock(&philo->argument->end_mutex);
-	if (philo->argument->stop_simulation == 0)
-	{
-		philo->argument->stop_simulation = 1;
-		print_time_stamp_with_message(philo, "died");
-	}
-	pthread_mutex_unlock(&philo->argument->end_mutex);
-	if (philo->is_holding_forks)
-		put_forks(philo, 1);
-	pthread_exit(NULL);
-}
-
-void	check_eating_limits(t_philosophers *philo)
-{
-	if (philo->argument->times_must_eat != -1
-		&& philo->total_eat_count >= philo->argument->times_must_eat)
-	{
-		pthread_mutex_lock(&philo->argument->end_mutex);
-		if (philo->argument->stop_simulation == 0)
-			philo->argument->stop_simulation = 1;
-		pthread_mutex_unlock(&philo->argument->end_mutex);
-	}
-}
 
 void	check_live_or_die(t_philosophers *philo)
 {
@@ -61,49 +23,6 @@ void	check_live_or_die(t_philosophers *philo)
 		pthread_exit(NULL);
 	}
 	pthread_mutex_unlock(&philo->argument->end_mutex);
-}
-
-void	start_eating_timer(struct timeval *start_behaivor)
-{
-	if (gettimeofday(start_behaivor, NULL) != 0)
-		exit_with_message("Error: gettimeofday failed");
-}
-
-double	calculate_elapsed_time(struct timeval *start, struct timeval *now)
-{
-	int	milliseconds;
-	int	useconds;
-
-	milliseconds = (now->tv_sec - start->tv_sec) * 1000;
-	useconds = (now->tv_usec - start->tv_usec) / 1000;
-	return (milliseconds + useconds);
-}
-
-void	perform_eating(t_philosophers *philo, struct timeval start_behaivor)
-{
-	double	elapsed;
-
-	elapsed = 0;
-	print_time_stamp_with_message(philo, "is eating");
-	while (elapsed <= philo->argument->time_to_eat)
-	{
-		if (gettimeofday(&philo->now, NULL) != 0)
-			exit_with_message("Error: gettimeofday failed");
-		elapsed = calculate_elapsed_time(&start_behaivor, &philo->now);
-		usleep(500);
-		check_live_or_die(philo);
-	}
-}
-
-void	eat(t_philosophers *philo)
-{
-	struct timeval	start_behaivor;
-
-	start_eating_timer(&start_behaivor);
-	perform_eating(philo, start_behaivor);
-	if (gettimeofday(&philo->start, NULL) != 0)
-		exit_with_message("Error: gettimeofday failed");
-	philo->total_eat_count++;
 }
 
 void	put_forks(t_philosophers *philo, int put_end)
@@ -169,31 +88,4 @@ void	take_forks(t_philosophers *philo)
 	print_time_stamp_with_message(philo, "has taken a fork");
 	eat(philo);
 	put_forks(philo, 0);
-}
-
-void	sleeping(t_philosophers *philo)
-{
-	struct timeval	start_behaivor;
-	int				milliseconds;
-	int				useconds;
-	double			elapsed;
-
-	elapsed = 0;
-	if (gettimeofday(&start_behaivor, NULL) != 0)
-	{
-		exit_with_message("Error: gettimeofday failed");
-	}
-	print_time_stamp_with_message(philo, "is sleeping");
-	while (elapsed <= philo->argument->time_to_sleep)
-	{
-		if (gettimeofday(&philo->now, NULL) != 0)
-		{
-			exit_with_message("Error: gettimeofday failed");
-		}
-		milliseconds = (philo->now.tv_sec - start_behaivor.tv_sec) * 1000;
-		useconds = (philo->now.tv_usec - start_behaivor.tv_usec) / 1000;
-		elapsed = milliseconds + useconds;
-		usleep(500);
-		check_live_or_die(philo);
-	}
 }
